[test] Convert initMatchers to TypeScript
