+++ test/karma.conf.js
@@ -55,7 +55,7 @@ module.exports = function setKarmaConfig(config) {
       module: {
         rules: [
           {
-            test: /\.js$/,
+            test: /\.(js|ts)$/,
             loader: 'babel-loader',
             exclude: /node_modules/,
           },
@@ -80,6 +80,7 @@ module.exports = function setKarmaConfig(config) {
           '@testing-library/react/pure':
             '@testing-library/react/dist/@testing-library/react.pure.esm',
         },
+        extensions: ['.js', '.ts'],
       },
     },
     webpackMiddleware: {
