+++ test/utils/index.js
@@ -6,3 +6,4 @@ export { default as createServerRender } from './createServerRender';
 export { default as createShallow } from './createShallow';
 export { default as findOutermostIntrinsic, wrapsIntrinsicElement } from './findOutermostIntrinsic';
 export { default as getClasses } from './getClasses';
+export {} from './initMatchers';
