+++ test/utils/setupBabel.js
@@ -1,0 +1,3 @@
+require('@babel/register')({
+  extensions: ['.js', '.ts'],
+});
